backend/requirements.txt
# web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# database
supabase==2.0.3

# pdf processing
pypdf==3.17.1

# ai/ml
anthropic==0.7.8

# environment management
python-dotenv==1.0.0

# additional utilities (if needed)
python-multipart==0.0.6

railtracks 
railtracks-cli